@@ -37,46 +37,44 @@
     "cf-core": "4.10.0",
     "cf-layout": "5.0.0",
     "cf-typography": "5.0.1",
-    "core-js": "2.5.7",
     "highcharts": "https://github.com/cfpb/highcharts-dist#fix-missing-plotline-label",
-    "xdr": "https://github.com/contolini/xdr"
+    "xdr": "0.5.3"
   },
   "devDependencies": {
+    "@babel/core": "7.1.2",
+    "@babel/preset-env": "^7.0.0",
     "autoprefixer": "9.1.5",
-    "babel-core": "7.0.0-beta.3",
+    "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "23.6.0",
-    "babel-loader": "7.1.5",
-    "babel-preset-env": "^7.0.0-beta.3",
-    "browser-sync": "2.24.6",
+    "babel-loader": "8.0.4",
+    "browser-sync": "2.26.0",
     "compression": "1.7.3",
     "cors": "2.8.4",
     "del": "3.0.0",
     "express": "4.16.3",
     "gulp": "4.0.0",
     "gulp-changed": "3.2.0",
-    "gulp-connect": "5.5.0",
-    "gulp-cssmin": "0.2.0",
+    "gulp-connect": "5.6.1",
     "gulp-eslint": "5.0.0",
     "gulp-header": "2.0.5",
     "gulp-imagemin": "4.1.0",
-    "gulp-less": "3.4.0",
+    "gulp-less": "4.0.1",
     "gulp-notify": "3.2.0",
     "gulp-postcss": "8.0.0",
     "gulp-rename": "1.4.0",
     "gulp-sourcemaps": "2.6.4",
     "jest": "23.6.0",
     "jest-cli": "23.6.0",
-    "less": "2.7.2",
+    "less": "3.8.1",
     "minimist": "1.2.0",
-    "postcss-unmq": "1.0.2",
     "request": "2.88.0",
-    "require-dir": "1.0.0",
+    "require-dir": "1.1.0",
     "sauce-connect-launcher": "1.2.4",
     "serve-static": "1.13.2",
     "through2": "2.0.3",
     "uglifyjs-webpack-plugin": "1.3.0",
     "vinyl-named": "1.1.0",
-    "webpack": "4.16.5",
-    "webpack-stream": "5.1.0"
+    "webpack": "4.20.2",
+    "webpack-stream": "5.1.1"
   }
 }
