@@ -1,7 +1,9 @@
+import accessibility from 'highcharts/js/modules/accessibility';
+import Highcharts from 'highcharts/js/highstock';
 import { getFirstNumber } from '../utils/calculation';
-const Highcharts = require( 'highcharts/js/highstock' );
-const process = require( '../utils/process-json' );
-require( 'highcharts/js/modules/accessibility' )( Highcharts );
+import * as process from '../utils/process-json';
+
+accessibility( Highcharts );
 
 Highcharts.setOptions( {
   lang: {
@@ -61,7 +63,7 @@ function _getTickValue( value ) {
 
 class LineChart {
   constructor( { el, description, data, metadata, yAxisLabel } ) {
-    data = process.originations( data[0], metadata );
+    data = process.processNumOriginationsData( data[0], metadata );
 
     const options = {
       chart: {
@@ -235,4 +237,4 @@ class LineChart {
   }
 }
 
-module.exports = LineChart;
+export default LineChart;
