@@ -4,16 +4,13 @@ const configPkg = config.pkg;
 const configBanner = config.banner;
 const handleErrors = require( '../utils/handle-errors' );
 const gulp = require( 'gulp' );
-const gulpCssmin = require( 'gulp-cssmin' );
 const gulpHeader = require( 'gulp-header' );
 const gulpLess = require( 'gulp-less' );
 const gulpPostcss = require( 'gulp-postcss' );
 const gulpSourcemaps = require( 'gulp-sourcemaps' );
 const gulpRename = require( 'gulp-rename' );
-const postcssUnmq = require( 'postcss-unmq' );
 
 /**
- * Process legacy CSS for IE9 only.
  * @returns {PassThrough} A source stream.
  */
 function stylesDemo() {
@@ -24,59 +21,14 @@ function stylesDemo() {
     .pipe( gulpPostcss( [
       autoprefixer( {
         browsers: [
-          'last 2 version',
-          'not ie <= 8',
-          'android 4',
-          'BlackBerry 7',
-          'BlackBerry 10' ]
+          'last 2 version'
+        ]
       } )
     ] ) )
     .pipe( gulpHeader( configBanner, { pkg: configPkg } ) )
     .pipe( gulpSourcemaps.write( '.' ) )
     .pipe( gulp.dest( config.demoStyles.dest ) );
 }
-
-/**
- * Process legacy CSS for IE9 only.
- * @returns {PassThrough} A source stream.
- */
-function demoStylesIE9() {
-  return gulp.src( config.demoStyles.cwd + config.demoStyles.src )
-    .pipe( gulpLess( config.demoStyles.settings ) )
-    .on( 'error', handleErrors )
-    .pipe( gulpPostcss( [
-      autoprefixer( {
-        browsers: [ 'ie 9' ]
-      } )
-    ] ) )
-    .pipe( gulpCssmin() )
-    .pipe( gulpRename( {
-      suffix:  '.ie9',
-      extname: '.css'
-    } ) )
-    .pipe( gulp.dest( config.demoStyles.dest ) );
-}
-
-gulp.task( 'styles:demoStylesIE8', () => {
-  const stream = gulp.src( config.demoStyles.cwd + config.demoStyles.src )
-    .pipe( gulpLess( config.demoStyles.settings ) )
-    .on( 'error', handleErrors )
-    .pipe( gulpPostcss( [
-      postcssUnmq( {
-        width: '75em'
-      } ),
-      autoprefixer( {
-        browsers: [ 'ie 7-8' ]
-      } )
-    ] ) )
-    .pipe( gulpCssmin() )
-    .pipe( gulpRename( {
-      suffix:  '.ie8',
-      extname: '.css'
-    } ) )
-    .pipe( gulp.dest( config.demoStyles.dest ) );
-  return stream;
-} );
 
 gulp.task( 'styles:chartsConcat', () => {
   const stream = gulp.src( config.styles.cwd + config.styles.src )
@@ -89,11 +41,8 @@ gulp.task( 'styles:chartsConcat', () => {
     .pipe( gulpPostcss( [
       autoprefixer( {
         browsers: [
-          'last 2 version',
-          'not ie <= 8',
-          'android 4',
-          'BlackBerry 7',
-          'BlackBerry 10' ]
+          'last 2 version'
+        ]
       } )
     ] ) )
     .pipe( gulp.dest( config.styles.dest ) );
@@ -108,11 +57,8 @@ gulp.task( 'styles:chartsMinify', () => {
     .pipe( gulpPostcss( [
       autoprefixer( {
         browsers: [
-          'last 2 version',
-          'not ie <= 8',
-          'android 4',
-          'BlackBerry 7',
-          'BlackBerry 10' ]
+          'last 2 version'
+        ]
       } )
     ] ) )
     .pipe( gulpRename( {
@@ -123,14 +69,6 @@ gulp.task( 'styles:chartsMinify', () => {
 } );
 
 gulp.task( 'styles:demo', stylesDemo );
-gulp.task( 'styles:demoStylesIE9', demoStylesIE9 );
-
-gulp.task( 'styles:demoIE',
-  gulp.series(
-    'styles:demoStylesIE8',
-    'styles:demoStylesIE9'
-  )
-);
 
 gulp.task( 'styles:charts',
   gulp.series(
@@ -142,7 +80,6 @@ gulp.task( 'styles:charts',
 gulp.task( 'styles',
   gulp.series(
     'styles:demo',
-    'styles:demoIE',
     'styles:charts'
   )
 );
