@@ -104,8 +104,8 @@ function _getStorageType( storage ) {
 }
 
 // Expose public methods.
-module.exports = {
-  setItem:    setItem,
-  getItem:    getItem,
-  removeItem: removeItem
+export {
+  setItem,
+  getItem,
+  removeItem
 };
